+++ macros/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- fielded variants and `{field}` interpolation in extras (Formatted)
+
 ## [0.4.0](https://github.com/Cardosaum/cognomen/releases/tag/cognomen-macros-v0.4.0) - 2026-08-16
 
 ### Other